Fixed icon

+++ Judge.java
@@ -23,28 +23,28 @@ class Judge{
 		verdictWindow = new JFrame();
 		verdictWindow.setVisible(false);
 	}
-		
+
 	public void judgeProblem(int problem, String language, String code){
 		AutoJudge.gui.disableJudge();
 		Bundle params = new Bundle(language, code, AutoJudge.problemHandler.getProblems().get(problem));
-		
+
 		AutoJudge.runWindow.reset(AutoJudge.problemHandler.getProblems().get(problem));
 		try{
 			File tmpf = new File("submissions\\");
 			tmpf.mkdirs();
-			
+
 			FileWriter fw = new FileWriter(new File(params.saveFile));
 			fw.write(code);
 			fw.flush();
 			fw.close();
-			
+
 			new CompileThread(params).start();
 		}catch(Exception e){
 			AutoJudge.gui.enableJudge();
 			JOptionPane.showMessageDialog(null, "Failed to judge run");
 		}
 	}
-	
+
 	public boolean checkForCompileErrorOverride(String error){
 		DialogBox d = new DialogBox(null, "Compile Error", "The given program failed to compile properly. Do you want to continue?", "Continue");
 		JTextArea errorText = HelperLib.getTextArea(error);
@@ -52,7 +52,7 @@ class Judge{
 		d.setContents(errorText);
 		return d.showToUser();
 	}
-	
+
 	public void showLastBundle(){
 		if(lastBundle.terminated != Integer.MAX_VALUE){
 			giveRunTerminatedVerdict(lastBundle);
@@ -60,11 +60,11 @@ class Judge{
 			giveVerdict(lastBundle);
 		}
 	}
-	
+
 	public void giveVerdict(Bundle b){
 		AutoJudge.runWindow.disableTerminate();
 		lastBundle = b;
-		
+
 		if(b.problem.checkerFile != null){
 			if(b.runtimeError < Math.min(Math.min(b.timeLimitExceeded, b.checkerRuntimeError), Math.min(b.wrongAnswer, b.outputFormatError))){
 				giveCasesDisplayedVerdict("Runtime Error in test case "+(b.runtimeError+1), b.runtimeError, 3, new String[]{"Input:", "Runtime Error:", "Checker Notes:"}, new ArrayList[]{b.inputs, b.errors, b.checkerNotes});
@@ -92,15 +92,15 @@ class Judge{
 				giveCasesDisplayedVerdict("Accepted", 0, 3, new String[]{"Input:", "Team's Output:", "Judge's Output:"}, new ArrayList[]{b.inputs, b.teamOutputs, b.judgeOutputs});
 			}
 		}
-		
+
 		AutoJudge.gui.enableJudge();
 	}
-	
+
 	public void giveRunTerminatedVerdict(Bundle b){
 		giveCasesDisplayedVerdict("Terminated", b.terminated, 3, new String[]{"Input:", "Team's Output:", "Judge's Output:"}, new ArrayList[]{b.inputs, b.teamOutputs, b.judgeOutputs});
 		AutoJudge.gui.enableJudge();
 	}
-	
+
 	public boolean checkWrongAnswer(String a, String b,Problem p){
 		if(p.usePrecisionChecker){
 			return RunChecker.judge(a,b,p.precisionExponent,"Tolerate all whitespace");
@@ -119,7 +119,7 @@ class Judge{
 			return !a.equals(b);
 		}
 	}
-	
+
 	public void giveNoCasesVerdict(String type, String message){
 		AutoJudge.runWindow.disableTerminate();
 		JTextArea tmp = HelperLib.getTextArea(message);
@@ -129,20 +129,20 @@ class Judge{
 		JOptionPane.showMessageDialog(null, scroll, type, JOptionPane.PLAIN_MESSAGE);
 		AutoJudge.gui.enableJudge();
 	}
-	
+
 	public void giveCasesDisplayedVerdict(String type, int i, int cols, String[] labels, ArrayList<String>[] strings){
 		verdictWindow.setVisible(false);
 		verdictWindow.getContentPane().removeAll();
-		
+
 		JPanel mainPanel = new JPanel();
 		mainPanel.setLayout(new BorderLayout());
-		
+
 		JLabel display = new JLabel("Currently Displaying Test Case "+(i+1));
 		mainPanel.add(display, BorderLayout.NORTH);
-		
+
 		JPanel compPanel = new JPanel();
 		compPanel.setLayout(new GridLayout(1, cols));
-		
+
 		JTextArea[] lines = new JTextArea[cols];
 		JTextArea[] text = new JTextArea[cols];
 		for(int cur=0; cur<cols; cur++){
@@ -166,15 +166,15 @@ class Judge{
 			panel.add(scroll, BorderLayout.CENTER);
 			compPanel.add(panel);
 		}
-		
+
 		mainPanel.add(compPanel, BorderLayout.CENTER);
-		
+
 		JPanel bottomPanel = new JPanel();
 		bottomPanel.setLayout(new BorderLayout());
-		
+
 		JPanel changeCase = new JPanel();
 		changeCase.setLayout(new FlowLayout());
-		
+
 		JLabel testCase = new JLabel("Select Test Case: ");
 		changeCase.add(testCase);
 		class CircularNumberModel extends SpinnerNumberModel {
@@ -186,12 +186,12 @@ class Judge{
 			}
 
 			public Object getNextValue() {
-				if(getValue() == max) return min;
+				if((int)getValue() == max) return min;
 				return (int)getValue()+1;
 			}
 
 			public Object getPreviousValue() {
-				if(getValue() == min) return max;
+				if((int)getValue() == min) return max;
 				return (int)getValue()-1;
 			}
 		}
@@ -210,11 +210,11 @@ class Judge{
 				this.text = text;
 				this.strings = strings;
 			}
-			
+
 			public void stateChanged(ChangeEvent e){
 				JSpinner spinner = (JSpinner)e.getSource();
 				SpinnerModel model = spinner.getModel();
-				
+
 				int index = (int)model.getValue();
 				for(int i=0; i<cols; i++){
 					lines[i].setText(HelperLib.getLines(strings[i].get(index-1)));
@@ -229,7 +229,7 @@ class Judge{
 			public SpinnerMouseWheelListener(CircularNumberModel model){
 				this.model = model;
 			}
-			
+
 			public void mouseWheelMoved(MouseWheelEvent e){
 				int mov = e.getWheelRotation();
 				while(mov < 0){
@@ -248,12 +248,12 @@ class Judge{
 			public SpinnerKeyListener(CircularNumberModel model){
 				this.model = model;
 			}
-			
+
 			public void keyPressed(KeyEvent e){
 				switch(e.getKeyCode()){
-					case KeyEvent.VK_UP: 
+					case KeyEvent.VK_UP:
 						model.setValue(model.getNextValue()); break;
-					case KeyEvent.VK_DOWN: 
+					case KeyEvent.VK_DOWN:
 						model.setValue(model.getPreviousValue()); break;
 				}
 			}
@@ -263,7 +263,7 @@ class Judge{
 		//mainPanel.addKeyListener(new SpinnerKeyListener(spinnerModel)); --- y u no work :<
 		changeCase.add(spinner);
 		bottomPanel.add(changeCase, BorderLayout.CENTER);
-		
+
 		JButton closeButton = new JButton("Close");
 		class CloseButtonListener implements ActionListener{
 			JFrame parent;
@@ -276,11 +276,11 @@ class Judge{
 		}
 		closeButton.addActionListener(new CloseButtonListener(verdictWindow));
 		bottomPanel.add(closeButton, BorderLayout.SOUTH);
-		
+
 		mainPanel.add(bottomPanel, BorderLayout.SOUTH);
-		
+
 		compPanel.setPreferredSize(new Dimension(850, 300));
-		
+
 		verdictWindow.add(mainPanel);
 		verdictWindow.setTitle(type);
 		verdictWindow.pack();
